pyproject.toml
[project]
name = "reno-structured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jan Scholz", email = "jan.scholz@gmail.com" }
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
    "reno>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
]
